+++ src/v/kafka/client/test/fetch.cc
@@ -11,6 +11,7 @@
 
 #include "kafka/protocol/fetch.h"
 
+#include "coproc/reflection.h"
 #include "kafka/client/client.h"
 #include "kafka/client/configuration.h"
 #include "kafka/client/test/fixture.h"
