+++ src/v/kafka/client/test/produce.cc
@@ -9,6 +9,7 @@
 
 #include "kafka/protocol/produce.h"
 
+#include "coproc/reflection.h"
 #include "kafka/client/client.h"
 #include "kafka/client/configuration.h"
 #include "kafka/client/test/fixture.h"
