+++ src/v/kafka/server/tests/fetch_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "coproc/reflection.h"
 #include "kafka/protocol/batch_consumer.h"
 #include "kafka/server/handlers/fetch.h"
 #include "kafka/types.h"
