+++ src/v/kafka/server/tests/metadata_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "coproc/reflection.h"
 #include "kafka/protocol/errors.h"
 #include "librdkafka/rdkafkacpp.h"
 #include "model/fundamental.h"
