+++ src/v/kafka/server/tests/topic_recreate_test.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "coproc/reflection.h"
 #include "kafka/protocol/batch_consumer.h"
 #include "kafka/protocol/delete_topics.h"
 #include "kafka/protocol/errors.h"
