+++ src/v/pandaproxy/test/fetch.cc
@@ -7,6 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
+#include "coproc/reflection.h"
 #include "http/client.h"
 #include "pandaproxy/configuration.h"
 #include "pandaproxy/test/pandaproxy_fixture.h"
